Bugs with types in random are fixed

+++ testlib/tree.hpp
@@ -168,7 +168,7 @@ public:
 		for (size_t i = 1; i < numberVertices; i++)
 			edges[i - 1] = std::make_pair(i, i - 1);
 		tree::Tree tree = tree::Tree(edges);
-		tree = tree.rehang(rnd.next<int>(0, numberVertices - 1));
+		tree = tree.rehang(rnd.next<size_t>(0, numberVertices - 1));
 		tree = tree::shuffle(tree, rnd);
 		return tree;
 	}
@@ -218,7 +218,7 @@ public:
 		for (size_t i = 1; i < numberVertices; i++)
 			edges[i - 1] = std::make_pair(i , (i - 1) / vertexDegree);
 		tree::Tree tree = tree::Tree(edges);
-		tree = tree.rehang(rnd.next<int>(0, numberVertices - 1));
+		tree = tree.rehang(rnd.next<size_t>(0, numberVertices - 1));
 		tree = tree::shuffle(tree, rnd);
 		return tree;
 	}
@@ -228,13 +228,13 @@ class FluffyBambooGenerator : public Generator<tree::Tree> {
 public:
 	tree::Tree generate(Random& rnd, size_t numberVertices) const {
 		std::vector<std::pair<size_t, size_t> > edges(numberVertices - 1);
-		size_t bambooSize = rnd.next<int>(1, numberVertices);
+		size_t bambooSize = rnd.next<size_t>(1, numberVertices);
 		for (size_t i = 1; i < bambooSize; i++)
 			edges[i - 1] = std::make_pair(i, i - 1);
 		for (size_t i = bambooSize; i < numberVertices; i++)
-			edges[i - 1] = std::make_pair(i, rnd.next<int>(0, bambooSize - 1));
+			edges[i - 1] = std::make_pair(i, rnd.next<size_t>(0, bambooSize - 1));
 		tree::Tree tree = tree::Tree(edges);
-		tree = tree.rehang(rnd.next<int>(0, numberVertices - 1));
+		tree = tree.rehang(rnd.next<size_t>(0, numberVertices - 1));
 		tree = tree::shuffle(tree, rnd);
 		return tree;
 	}
